+++ crates/bitwarden-wasm-internal/src/platform/token_provider.rs
@@ -1,0 +1,43 @@
+use bitwarden_core::client::internal::ClientManagedTokens;
+use bitwarden_threading::ThreadBoundRunner;
+use wasm_bindgen::{prelude::wasm_bindgen, JsValue};
+
+#[wasm_bindgen(typescript_custom_section)]
+const TOKEN_CUSTOM_TS_TYPE: &'static str = r#"
+export interface TokenProvider {
+    get_access_token(): Promise<string>;
+}
+"#;
+
+#[wasm_bindgen]
+extern "C" {
+    #[wasm_bindgen(js_name = TokenProvider)]
+    pub type JsTokenProvider;
+
+    #[wasm_bindgen(method)]
+    pub async fn get_access_token(this: &JsTokenProvider) -> JsValue;
+}
+
+pub(crate) struct WasmClientManagedTokens(ThreadBoundRunner<JsTokenProvider>);
+
+impl WasmClientManagedTokens {
+    pub fn new(js_provider: JsTokenProvider) -> Self {
+        Self(ThreadBoundRunner::new(js_provider))
+    }
+}
+
+impl std::fmt::Debug for WasmClientManagedTokens {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("WasmClientManagedTokens").finish()
+    }
+}
+
+#[async_trait::async_trait]
+impl ClientManagedTokens for WasmClientManagedTokens {
+    async fn get_access_token(&self) -> Option<String> {
+        self.0
+            .run_in_thread(async move |c| c.get_access_token().await.as_string())
+            .await
+            .unwrap_or_default()
+    }
+}
